\name{Metamaps2-package}
\alias{Metamaps2-package}
\alias{Metamaps2}
\docType{package}
\title{METAMAPS: Toolset for exploring metagenomic data and analyses}
\description{Useful for building Metabolic Pathways and Querying Metamaps graphdatabase}
\details{
\tabular{ll}{Package: \tab Metamaps2\cr
Type: \tab Package\cr
Title: \tab METAMAPS: Toolset for exploring metagenomic data and analyses\cr
Version: \tab 2.2.9\cr
Date: \tab 2013-04-27\cr
Author: \tab Wesley Chin Lui GOI\cr
Maintainer: \tab Wesley Goi <picy2k@gmail.com>\cr
License: \tab GPLv2\cr
LazyLoad: \tab yes\cr
Packaged: \tab 2013-07-24 06:36:10 UTC; uesu\cr
Imports: \tab igraph,
RJSONIO,
RCurl\cr}
}
\author{Wesley Chin Lui GOI}

\keyword{ package }


